# py-genai/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Hello-GenAI API",
    "description": "REST API for the Hello-GenAI local chat application powered by Docker Model Runner.",
    "version": "2.0.0"
  },
  "servers": [
    { "url": "/", "description": "Current server" }
  ],
...
  "components": {
    "parameters": {
      "session_id": {
        "name": "session_id",
        "in": "path",
        "required": true,
        "description": "UUID of the session",
        "schema": { "type": "string", "format": "uuid" },
        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "Message is required" }
        }
      },
      "Ok": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean", "example": true }
        }
      },
...
      "Session": {
        "type": "object",
        "example": {
          "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "title": "How does Docker work?",
          "system_prompt": null,
          "model": "docker.io/ai/gemma3n:latest",
          "pinned": 0,
          "created_at": "2024-01-01T00:00:00+00:00",
          "updated_at": "2024-01-01T00:01:00+00:00"
        },
        "properties": {
          "id":            { "type": "string",  "format": "uuid" },
          "title":         { "type": "string",  "example": "How does Docker work?" },
          "system_prompt": { "type": "string",  "nullable": true },
          "model":         { "type": "string",  "nullable": true, "description": "Model used when this session was first created", "example": "docker.io/ai/gemma3n:latest" },
          "pinned":        { "type": "integer", "enum": [0, 1] },
          "created_at":    { "type": "string",  "format": "date-time" },
          "updated_at":    { "type": "string",  "format": "date-time" }
        }
      },
...
      "Preset": {
        "type": "object",
        "example": { "id": 1, "name": "Concise", "text": "You are a concise assistant.", "created_at": "2024-01-01T00:00:00+00:00" },
        "properties": {
          "id":         { "type": "integer" },
          "name":       { "type": "string", "maxLength": 60 },
          "text":       { "type": "string", "maxLength": 2000 },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Memory": {
        "type": "object",
        "example": {
          "id": 1,
          "content": "User prefers concise answers with code examples",
          "source_session_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "enabled": 1,
          "created_at": "2024-01-01T00:00:00+00:00",
          "updated_at": "2024-01-01T00:00:00+00:00"
        },
        "properties": {
          "id":                { "type": "integer" },
          "content":           { "type": "string", "maxLength": 300 },
          "source_session_id": { "type": "string", "format": "uuid", "nullable": true, "description": "Session the fact was extracted from; null for manually added memories" },
          "enabled":           { "type": "integer", "enum": [0, 1], "description": "Disabled memories stay stored but are not recalled" },
          "created_at":        { "type": "string", "format": "date-time" },
          "updated_at":        { "type": "string", "format": "date-time" }
        }
      },
...
      "ImportRequest": {
        "type": "object",
        "required": ["messages"],
        "example": {
          "title": "Imported Chat",
          "system_prompt": null,
          "messages": [
            { "role": "user",      "content": "Hello!" },
            { "role": "assistant", "content": "Hi there! How can I help?" }
          ]
        },
        "properties": {
          "title":         { "type": "string", "example": "Imported Chat" },
          "system_prompt": { "type": "string", "nullable": true },
          "messages": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "role":    { "type": "string", "enum": ["user", "assistant"] },
                "content": { "type": "string" }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/presets": {
      "get": {
        "tags": ["Presets"],
        "summary": "List system prompt presets",
        "responses": {
          "200": {
            "description": "Presets ordered by name",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Preset" } }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Presets"],
        "summary": "Create a preset",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "text"],
                "properties": {
                  "name": { "type": "string", "maxLength": 60, "example": "Concise" },
                  "text": { "type": "string", "maxLength": 2000, "example": "You are a concise assistant." }
                }
              },
              "example": { "name": "Concise", "text": "You are a concise assistant." }
            }
          }
        },
        "responses": {
          "201": { "description": "Preset created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Preset" } } } },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/api/sessions": {
      "get": {
        "tags": ["Sessions"],
        "summary": "List all sessions",
        "description": "Returns all sessions ordered by pinned status then last updated.",
        "responses": {
          "200": {
            "description": "Array of sessions",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Session" } }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Sessions"],
        "summary": "Create a session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title":         { "type": "string", "example": "New Chat" },
                  "system_prompt": { "type": "string", "nullable": true, "example": null }
                }
              },
              "example": {
                "title": "New Chat",
                "system_prompt": null
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Session created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": { "session_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6" },
                  "properties": {
                    "session_id": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/import": {
      "post": {
        "tags": ["Sessions"],
        "summary": "Import a conversation from JSON",
        "description": "Creates a new session and bulk-inserts the provided messages. Useful for restoring exported chats.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ImportRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Session created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": { "session_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6" },
                  "properties": {
                    "session_id": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/api/sessions/{session_id}/export": {
      "get": {
        "tags": ["Sessions"],
        "summary": "Export session as Markdown or JSON",
        "description": "Downloads the full conversation. The default format is Markdown; `format=json` produces a file that round-trips through `POST /api/sessions/import`.",
        "parameters": [
          { "$ref": "#/components/parameters/session_id" },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "description": "Export format: `md` (default) or `json` (re-importable)",
            "schema": { "type": "string", "enum": ["md", "json"], "default": "md" },
            "example": "json"
          }
        ],
        "responses": {
          "200": {
            "description": "File download",
            "content": {
              "text/markdown": { "schema": { "type": "string" } },
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "title": "Docker basics",
                    "system_prompt": null,
                    "model": "docker.io/ai/gemma4:latest",
                    "exported_at": "2024-01-01T00:01:00+00:00",
                    "messages": [
                      { "role": "user", "content": "What is a container?" },
                      { "role": "assistant", "content": "A container is…" }
                    ]
                  }
                }
              }
            }
          },
          "404": { "description": "Session not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/api/memories": {
      "get": {
        "tags": ["Memory"],
        "summary": "List memories",
        "description": "All stored facts about the user, including disabled ones.",
        "responses": {
          "200": {
            "description": "Memories in creation order",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Memory" } }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Memory"],
        "summary": "Add a memory manually",
        "description": "Duplicate content (case-insensitive) is not stored twice — the existing memory's id is returned.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["content"],
                "properties": {
                  "content": { "type": "string", "maxLength": 300, "example": "User's name is Anirban" }
                }
              },
              "example": { "content": "User's name is Anirban" }
            }
          }
        },
        "responses": {
          "201": { "description": "Memory stored", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Memory" } } } },
          "400": { "description": "Missing content", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "409": { "description": "Memory is full (MEMORY_MAX_ITEMS reached)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "delete": {
        "tags": ["Memory"],
        "summary": "Forget everything",
        "description": "Deletes all stored memories.",
        "responses": {
          "200": {
            "description": "Cleared",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": { "ok": true, "deleted": 7 },
                  "properties": {
                    "ok":      { "type": "boolean" },
                    "deleted": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/backup": {
      "get": {
        "tags": ["Backup"],
        "summary": "Export a full backup",
        "description": "Downloads every conversation (with messages) and all presets as a single JSON file that can be restored via POST.",
        "responses": {
          "200": {
            "description": "Backup file",
            "content": { "application/json": { "schema": { "type": "object" } } }
          }
        }
      },
      "post": {
        "tags": ["Backup"],
        "summary": "Restore a full backup",
        "description": "Recreates sessions, messages, and presets from a backup produced by GET /api/backup.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["sessions"],
                "properties": {
                  "sessions": { "type": "array", "items": { "type": "object" } },
                  "presets":  { "type": "array", "items": { "type": "object" } }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Restored",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": { "imported_sessions": 12 },
                  "properties": { "imported_sessions": { "type": "integer" } }
                }
              }
            }
          },
          "400": { "description": "Invalid backup", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/api/documents": {
      "get": {
        "tags": ["Documents"], "summary": "List knowledge-base documents",
        "parameters": [ { "name": "project_id", "in": "query", "required": false, "schema": { "type": "integer" } } ],
        "responses": { "200": { "description": "Documents + embeddings_available flag" } }
      },
      "post": {
        "tags": ["Documents"], "summary": "Ingest a document",
        "description": "Multipart PDF/text file upload (`file`) or JSON `{ filename, text }`. Optional `project_id`. Chunked and embedded when embeddings are configured.",
        "requestBody": { "required": true, "content": {
          "multipart/form-data": { "schema": { "type": "object", "properties": { "file": { "type": "string", "format": "binary" }, "project_id": { "type": "integer" } } } },
          "application/json": { "schema": { "type": "object", "properties": { "filename": { "type": "string" }, "text": { "type": "string" }, "project_id": { "type": "integer" } } } }
        } },
        "responses": { "201": { "description": "Ingested (document_id, chunks, embedded)" }, "400": { "description": "No text found" } }
      }
    },
    "/api/documents/{document_id}": {
      "delete": {
        "tags": ["Documents"], "summary": "Remove a document",
        "parameters": [ { "name": "document_id", "in": "path", "required": true, "schema": { "type": "integer" } } ],
        "responses": { "200": { "description": "Deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Ok" } } } } }
      }
    },
    "/api/templates": {
      "get": { "tags": ["Templates"], "summary": "List slash-command templates", "responses": { "200": { "description": "Templates" } } },
      "post": {
        "tags": ["Templates"], "summary": "Create a template",
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "required": ["trigger", "content"],
          "properties": { "trigger": { "type": "string", "example": "tldr" }, "title": { "type": "string" }, "content": { "type": "string" } }
        } } } },
        "responses": { "201": { "description": "Created" }, "400": { "description": "Missing fields" } }
      }
    },
...
  }
}